Use just PopArgument instead of PrefixArgumentHack
