+++ src/prefix/argument/code_block.rs
@@ -44,71 +44,68 @@ impl std::fmt::Display for CodeBlock {
     }
 }
 
-impl<'a> PopArgument<'a> for CodeBlock {
-    type Err = MalformedCodeBlock;
+fn pop_from(args: &str) -> Result<(&str, CodeBlock), Option<MalformedCodeBlock>> {
+    let args = args.trim_start();
 
+    let rest;
+    let mut code_block = if let Some(code_block) = args.strip_prefix("```") {
+        let code_block_end = code_block.find("```").ok_or(Some(MalformedCodeBlock))?;
+        rest = &code_block[(code_block_end + 3)..];
+        let mut code_block = &code_block[..code_block_end];
+
+        // If a word is preceded directly by the three backticks and succeeded directly by a
+        // newline, it's interpreted as the code block language
+        let mut language = None;
+        if let Some(first_newline) = code_block.find('\n') {
+            if !code_block[..first_newline].contains(char::is_whitespace) {
+                language = Some(&code_block[..first_newline]);
+                code_block = &code_block[(first_newline + 1)..];
+            }
+        }
+
+        // Discord strips empty lines from start and end, but only if their really empty (even
+        // whitespace on the line cancels the stripping)
+        let code_block = code_block.trim_start_matches('\n').trim_end_matches('\n');
+
+        CodeBlock {
+            code: code_block.to_owned(),
+            language: language.map(|x| x.to_owned()),
+        }
+    } else if let Some(code_line) = args.strip_prefix('`') {
+        let code_line_end = code_line.find('`').ok_or(Some(MalformedCodeBlock))?;
+        rest = &code_line[(code_line_end + 1)..];
+        let code_line = &code_line[..code_line_end];
+
+        CodeBlock {
+            code: code_line.to_owned(),
+            language: None,
+        }
+    } else {
+        return Err(None);
+    };
+
+    // Empty codeblocks like `` are not rendered as codeblocks by Discord
+    if code_block.code.is_empty() {
+        Err(Some(MalformedCodeBlock))
+    } else {
+        // discord likes to insert hair spaces at the end of code blocks sometimes for no reason
+        code_block.code = code_block.code.trim_end_matches('\u{200a}').to_owned();
+
+        Ok((rest, code_block))
+    }
+}
+
+#[async_trait::async_trait]
+impl<'a> PopArgument<'a> for CodeBlock {
     /// Parse a single-line or multi-line code block. The output of `Self::code` should mirror what
     /// the official Discord client renders, and the output of `Self::language` should mirror the
     /// official Discord client's syntax highlighting, if existent.
-    ///
-    /// ```rust
-    /// # use poise::{CodeBlock, PopArgument as _};
-    /// assert_eq!(
-    ///     CodeBlock::pop_from("`hello world`").unwrap().1,
-    ///     CodeBlock { code: "hello world".into(), language: None },
-    /// );
-    /// assert_eq!(
-    ///     CodeBlock::pop_from("```rust\nprintln!(\"Hello world!\");\n```").unwrap().1,
-    ///     CodeBlock { code: "println!(\"Hello world!\");".into(), language: Some("rust".into()) },
-    /// );
-    /// ```
-    fn pop_from(args: &'a str) -> Result<(&'a str, Self), Option<Self::Err>> {
-        let rest;
-        let mut code_block = if let Some(code_block) = args.strip_prefix("```") {
-            let code_block_end = code_block.find("```").ok_or(Some(MalformedCodeBlock))?;
-            rest = &code_block[(code_block_end + 3)..];
-            let mut code_block = &code_block[..code_block_end];
-
-            // If a word is preceded directly by the three backticks and succeeded directly by a
-            // newline, it's interpreted as the code block language
-            let mut language = None;
-            if let Some(first_newline) = code_block.find('\n') {
-                if !code_block[..first_newline].contains(char::is_whitespace) {
-                    language = Some(&code_block[..first_newline]);
-                    code_block = &code_block[(first_newline + 1)..];
-                }
-            }
-
-            // Discord strips empty lines from start and end, but only if their really empty (even
-            // whitespace on the line cancels the stripping)
-            let code_block = code_block.trim_start_matches('\n').trim_end_matches('\n');
-
-            Self {
-                code: code_block.to_owned(),
-                language: language.map(|x| x.to_owned()),
-            }
-        } else if let Some(code_line) = args.strip_prefix('`') {
-            let code_line_end = code_line.find('`').ok_or(Some(MalformedCodeBlock))?;
-            rest = &code_line[(code_line_end + 1)..];
-            let code_line = &code_line[..code_line_end];
-
-            Self {
-                code: code_line.to_owned(),
-                language: None,
-            }
-        } else {
-            return Err(None);
-        };
-
-        // Empty codeblocks like `` are not rendered as codeblocks by Discord
-        if code_block.code.is_empty() {
-            Err(Some(MalformedCodeBlock))
-        } else {
-            // discord likes to insert hair spaces at the end of code blocks sometimes for no reason
-            code_block.code = code_block.code.trim_end_matches('\u{200a}').to_owned();
-
-            Ok((rest, code_block))
-        }
+    async fn pop_from(
+        args: &'a str,
+        _: &serenity::Context,
+        _: &serenity::Message,
+    ) -> Result<(&'a str, Self), (Box<dyn std::error::Error + Send + Sync>, Option<String>)> {
+        pop_from(args).map_err(|e| (e.map_or(crate::TooFewArguments.into(), |e| e.into()), None))
     }
 }
 
@@ -125,7 +122,7 @@ fn test_pop_code_block() {
         ("```rust\n\n\n\n\nhi\n\n\n\n```", "hi", Some("rust")),
     ] {
         assert_eq!(
-            CodeBlock::pop_from(string).unwrap().1,
+            pop_from(string).unwrap().1,
             CodeBlock {
                 code: code.into(),
                 language: language.map(|x| x.into())
@@ -133,7 +130,7 @@ fn test_pop_code_block() {
         );
     }
 
-    assert_eq!(CodeBlock::pop_from(""), Err(None));
-    assert_eq!(CodeBlock::pop_from("``"), Err(Some(MalformedCodeBlock)));
-    assert_eq!(CodeBlock::pop_from("``````"), Err(Some(MalformedCodeBlock)));
+    assert_eq!(pop_from(""), Err(None));
+    assert_eq!(pop_from("``"), Err(Some(MalformedCodeBlock)));
+    assert_eq!(pop_from("``````"), Err(Some(MalformedCodeBlock)));
 }
