+++ src/prefix/argument/macros.rs
@@ -13,7 +13,7 @@ macro_rules! _parse_prefix {
         (Option<$type:ty $(,)?>)
         $( $rest:tt )*
     ) => {
-        match (&PhantomData::<$type>).pop(&$args, $ctx, $msg).await {
+        match <$type>::pop_from(&$args, $ctx, $msg).await {
             Ok(($args, token)) => {
                 let token: Option<$type> = Some(token);
                 $crate::_parse_prefix!($ctx $msg $args => [ $error $($preamble)* token ] $($rest)* );
@@ -31,7 +31,7 @@ macro_rules! _parse_prefix {
     ) => {
         let token: Option<$type> = None;
         $crate::_parse_prefix!($ctx $msg $args => [ $error $($preamble)* token ] $($rest)* );
-        match (&PhantomData::<$type>).pop(&$args, $ctx, $msg).await {
+        match <$type>::pop_from(&$args, $ctx, $msg).await {
             Ok(($args, token)) => {
                 let token: Option<$type> = Some(token);
                 $crate::_parse_prefix!($ctx $msg $args => [ $error $($preamble)* token ] $($rest)* );
@@ -73,7 +73,7 @@ macro_rules! _parse_prefix {
 
         let mut running_args = $args.clone();
         loop {
-            match (&PhantomData::<$type>).pop(&running_args, $ctx, $msg).await {
+            match <$type>::pop_from(&running_args, $ctx, $msg).await {
                 Ok((popped_args, token)) => {
                     tokens.push(token);
                     token_rest_args.push(popped_args.clone());
@@ -119,7 +119,7 @@ macro_rules! _parse_prefix {
         (#[flag] $name:literal)
         $( $rest:tt )*
     ) => {
-        if let Ok(($args, token)) = (&PhantomData::<String>).pop(&$args, $ctx, $msg).await {
+        if let Ok(($args, token)) = String::pop_from(&$args, $ctx, $msg).await {
             if token.eq_ignore_ascii_case($name) {
                 $crate::_parse_prefix!($ctx $msg $args => [ $error $($preamble)* true ] $($rest)* );
             }
@@ -133,7 +133,7 @@ macro_rules! _parse_prefix {
         ($type:ty)
         $( $rest:tt )*
     ) => {
-        match (&PhantomData::<$type>).pop(&$args, $ctx, $msg).await {
+        match <$type>::pop_from(&$args, $ctx, $msg).await {
             Ok(($args, token)) => {
                 $crate::_parse_prefix!($ctx $msg $args => [ $error $($preamble)* token ] $($rest)* );
             },
@@ -200,8 +200,7 @@ macro_rules! parse_prefix_args {
         ( $($type:tt)* )
     ),* $(,)? ) => {
         async {
-            use std::marker::PhantomData;
-            use $crate::PrefixArgumentHack as _;
+            use $crate::PopArgument as _;
 
             let ctx = $ctx;
             let msg = $msg;
